feat: Make more aggressive discovery / offboarding
Try to discover devices as quickly as possible. Works best
with a limited IP range, but since we lower the concurrency,
it might work better in general.

Also update the action to use latest only if the branch is main.
Otherwise tag the container image with the specific github hash.

Signed-off-by: Anastassios Nanos <ananos@nubificus.co.uk>